Refactor: move arbitraries into separate file
