+++ src/test/e2e/services/playground-service.test.ts
@@ -1,5 +1,5 @@
 import { PlaygroundService } from '../../../lib/services/playground-service';
-import { generateFeatures } from '../../../lib/routes/admin-api/playground.test';
+import { generateClientFeatures } from '../../arbitraries.test';
 import { generate as generateContext } from '../../../lib/openapi/spec/sdk-context-schema.test';
 import fc from 'fast-check';
 import { createTestConfig } from '../../config/test-config';
@@ -52,7 +52,7 @@ describe('the playground service (e2e)', () => {
         await fc.assert(
             fc
                 .asyncProperty(
-                    generateFeatures({ minLength: 1 }),
+                    generateClientFeatures({ minLength: 1 }),
                     generateContext(),
                     async (toggles, context) => {
                         await Promise.all(
